fix: add proper health endpoint unit tests using Node's test framework
- Convert test file from .mts to .test.ts format (required by test runner)
- Use Node's built-in node:test module instead of vitest
- Adjust tests to handle test environment without real database connection
- Tests verify response structure, timestamp validity, and check inclusion
- All 156 unit tests pass (including 3 new health endpoint tests)
